Move the GraphQL query into it's own class
