+++ .gitignore
@@ -2,6 +2,7 @@
 node_modules
 app
 sessions
+out
 
 .env
 .pid
