+++ package.json
@@ -30,7 +30,6 @@
     "mkdirp": "^0.5.1",
     "moment": "^2.22.1",
     "node-sass": "^4.8.3",
-    "opn": "^5.3.0",
     "passport": "^0.4.0",
     "passport-auth0": "^0.6.1",
     "prismjs": "^1.14.0",
@@ -38,7 +37,11 @@
     "querystring": "^0.2.0",
     "request-promise": "^4.2.2",
     "sane": "^2.5.0",
-    "session-file-store": "^1.2.0",
+    "session-file-store": "^1.2.0"
+  },
+  "devDependencies": {
+    "opn": "^5.3.0",
+    "jsdoc": "^3.5.5",
     "standard": "^11.0.1"
   }
 }
